refactor social icons, add yt icon, downgrade nextra
