+++ package.json
@@ -34,8 +34,8 @@
         "framer-motion": "^10.18.0",
         "next": "^15.0.3",
         "next-auth": "^5.0.0-beta.25",
-        "nextra": "4.0.0-app-router.31",
-        "nextra-theme-docs": "4.0.0-app-router.31",
+        "nextra": "4.0.0-app-router.28",
+        "nextra-theme-docs": "4.0.0-app-router.28",
         "postcss": "8.4.47",
         "prettier": "^3.3.3",
         "prisma": "^5.21.1",
